[js] change routing for all apps
