+++ src/App.test.tsx
@@ -1,8 +1,8 @@
-import React from 'react';
-import { render } from '@testing-library/react';
-import { App } from './exodus/main';
+import React from "react";
+import { render } from "@testing-library/react";
+import { App } from "./exodus/oldMain";
 
-test('renders learn react link', () => {
+test("renders learn react link", () => {
   const { getByText } = render(<App />);
   const linkElement = getByText(/learn react/i);
   expect(linkElement).toBeInTheDocument();
