+++ src/exodus/startup.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { AppContext } from "./main";
+import { AppContext } from "./oldMain";
 
 export const AppFrame = ({ children }: { children: JSX.Element }) => (
   <AppContext.Provider value={null}>{children}</AppContext.Provider>
